revisi pending, reparasi, fast track, report, inspection list

+++ ASP.NET-TEFA/Controllers/InspectionListController.cs
@@ -7,6 +7,7 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using ASP.NET_TEFA.Models;
 using Newtonsoft.Json;
+using NuGet.Versioning;
 
 namespace ASP.NET_TEFA.Controllers
 {
@@ -24,7 +25,9 @@ namespace ASP.NET_TEFA.Controllers
         public async Task<IActionResult> Index(string idBooking)
         {
             // Mencari pemesanan berdasarkan ID
-            var trsBooking = await _context.TrsBookings.FindAsync(idBooking);
+            var trsBooking = await _context.TrsBookings
+            .Include(t => t.IdVehicleNavigation)
+            .FirstOrDefaultAsync(t => t.IdBooking == idBooking);
 
             // Jika pemesanan tidak ditemukan, kembalikan NotFound
             if (trsBooking == null)
@@ -48,6 +51,7 @@ namespace ASP.NET_TEFA.Controllers
 
             // Menyiapkan data untuk ditampilkan di View
             ViewBag.IdBooking = idBooking;
+            ViewBag.booking = trsBooking;
 
             return View(await trsInspectionLists);
         }
@@ -55,12 +59,13 @@ namespace ASP.NET_TEFA.Controllers
         // Membuat pemeriksaan baru berdasarkan input formulir
         [AuthorizedUser("HEAD MECHANIC")]
         [HttpPost]
-        [ValidateAntiForgeryToken]
         public IActionResult Create(IFormCollection form)
         {
             // Memeriksa pemesanan berdasarkan ID
             string idBooking = form["IdBooking"].ToString();
-            var booking = _context.TrsBookings.Find(idBooking);
+            var booking = _context.TrsBookings
+            .Include(t => t.IdVehicleNavigation)
+            .FirstOrDefault(t => t.IdBooking == idBooking);
 
             // Jika pemesanan tidak ditemukan, kembalikan NotFound
             if (booking == null)
@@ -75,49 +80,56 @@ namespace ASP.NET_TEFA.Controllers
                 return RedirectToAction("Index", "Reparation", new { idBooking = booking.IdBooking });
             }
 
-            // Menghapus data pemeriksaan lama terkait dengan pemesanan
-            var existingInspections = _context.TrsInspectionLists.Where(i => i.IdBooking == idBooking);
-            _context.TrsInspectionLists.RemoveRange(existingInspections);
-
-            // Menyimpan data pemeriksaan baru berdasarkan input yang diberikan
-            foreach (var key in form.Keys)
+            // Menghapus data pemeriksaan lama terkait dengan pemesanan jika kendaran mobil
+            if (booking.IdVehicleNavigation.Classify == "MOBIL")
             {
-                if (key.StartsWith("Checklist_"))
+                var existingInspections = _context.TrsInspectionLists.Where(i => i.IdBooking == idBooking);
+                _context.TrsInspectionLists.RemoveRange(existingInspections);
+
+                // Menyimpan data pemeriksaan baru berdasarkan input yang diberikan
+                foreach (var key in form.Keys)
                 {
-                    var idEquipment = key.Replace("Checklist_", "");
-                    var checklistValue = form[key].ToString();
-
-                    // Anda juga dapat memeriksa elemen deskripsi dengan nama yang sesuai di sini
-                    var descriptionKey = "Description_" + idEquipment;
-                    var descriptionValue = form[descriptionKey].ToString();
-
-                    // Konversi nilai checklist ke int (harus sesuai dengan tipe data Checklist di model)
-                    int checklist = int.Parse(checklistValue);
-
-                    // Membuat ID pemeriksaan secara acak
-                    Random random = new();
-                    string idInspection = random.Next(100000, 999999).ToString();
-
-                    // Membuat objek TrsInspectionList baru
-                    var data = new TrsInspectionList
+                    if (key.StartsWith("Checklist_"))
                     {
-                        IdInspection = idInspection,
-                        IdBooking = idBooking,
-                        IdEquipment = idEquipment.ToString(),
-                        Checklist = checklist,
-                        Description = descriptionValue.ToString()
-                    };
+                        var idEquipment = key.Replace("Checklist_", "");
+                        var checklistValue = form[key].ToString();
 
-                    // Menambahkan data pemeriksaan baru ke basis data
-                    _context.TrsInspectionLists.Add(data);
+                        // Anda juga dapat memeriksa elemen deskripsi dengan nama yang sesuai di sini
+                        var descriptionKey = "Description_" + idEquipment;
+                        var descriptionValue = form[descriptionKey].ToString();
+
+                        // Konversi nilai checklist ke int (harus sesuai dengan tipe data Checklist di model)
+                        int checklist = int.Parse(checklistValue);
+
+                        // Membuat ID pemeriksaan secara acak
+                        Random random = new();
+                        string idInspection = random.Next(100000, 999999).ToString();
+
+                        // Membuat objek TrsInspectionList baru
+                        var data = new TrsInspectionList
+                        {
+                            IdInspection = idInspection,
+                            IdBooking = idBooking,
+                            IdEquipment = idEquipment.ToString(),
+                            Checklist = checklist,
+                            Description = descriptionValue.ToString()
+                        };
+
+                        // Menambahkan data pemeriksaan baru ke basis data
+                        _context.TrsInspectionLists.Add(data);
+                    }
                 }
             }
 
             // Mengganti status reparasi dari "EKSEKUSI" menjadi "KONTROL"
             booking.RepairStatus = "EKSEKUSI";
 
-            // Menetapkan waktu mulai reparasi saat ini
-            booking.StartRepairTime = DateTime.Now;
+            // Menyimpan waktu mulai hanya sekali
+            if(booking.StartRepairTime == null)
+            {
+                // Menetapkan waktu mulai reparasi saat ini
+                booking.StartRepairTime = DateTime.Now;
+            }
 
             // Menetapkan kemajuan reparasi sebesar 45%
             booking.Progress = 45;
